package/gawk: don't sliently try autoreconf
Commit b5073d0ae35e (package/gawk: fix double free during kernel build)
backported an upstream patch, with hunks in autotools files, Makefile.am
and Makefile.in.

Such changes ususally require an explicitly autoreconf, but in this
case, Makefile.in is modified after Makefile,am, and a simple local test
did not trigger a silent autoreconf, so the patch was applied as-is.

That was a bit too optimitstic, though, and the build fails in some
situations.

Since the hunks in the autotools files are about tests, just drop those
hunks as we are not interested in running the tests.

Also drop the hunks about the changelogs while at it.

Fixes:
    http://autobuild.buildroot.org/results/068/0686697cb52d22c42ae2249db3442c008082fe72/
    http://autobuild.buildroot.org/results/754/75463639cb9cbc32108f059bd6044810a078db06/

Reported-by: Romain Naour <romain.naour@gmail.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>