package/modem-manager: bump version to 1.20.2
Switch package type to meson as the primary supported build system
from now on is meson and the autotools support as deprecated. The
package tarball is no longer supported, thus the MODEM_MANAGER_SOURCE
is changed to the gitlab archive URL.

All previously supported options are converted to the meson
configuraiton options and support for new ones is added. The new
options are: bash_completion, polkit, systemd, qrtr.

The dbus build dependency is needed since the meson build system uses
dbus pkg-config to determine install locations. The 'udevdir' has to
be specified also in case dbus is not installed. Modem-manager will
use its own built-in udev rule parser to determine modem specific
values.

For building with libqmi support version 1.32.2 is required and for
libmbim version 1.28.2 is required.

Signed-off-by: Heiko Thiery <heiko.thiery@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>