package/skeleton-systemd: move var factory comment closer to code
Commit 40c1d2c15815 (package/skeleton-systemd: systemd-ify mounting /var
tmpfs with ro rootfs) removed the fstab handling of /var, and left the
macro deal solely about keeping / read-only.

However, the code about how the var factory is handled was left above
the macro, which is now misleading and confusing.

Move the comment closer to the actual rootfs-pre-cmd hook which actually
deals with handling the var factory.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>