package/qemu: fix indentation for host-qemu sub-options in menuconfig
Commit 19446687c5dc (package/qemu: needs gcc >= 8) added a comment that
explains that host-qemu would not be available for gcc < 4.8.

However, it interspersed that comment between the main symbol and the
conditional block with the sub-options. This breaks the indentation of
the sub-options in menuconfig, where they appear at the same level as
the main symbol, rather than indented below it.

Fix that by moving the comment before the main symbol.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>