package/qemu: allow building only the tools
Commit fa7874307751 (package/qemu: refactor target emulator selection)
mistakenly forced to always build at least user or system emulation but
it should be possible to choose only the tools (e.g. for a guest VM
image).

Still, we don't want to configure qemu with neither system nor user
emulation, nor the tools enabled; at least one lust be enabled.

So, change the selection to force the tools to be enabled if neither the
system nor the user emulations are enabled.

Reinstate the the machine selection guard behind a dependency on whether
at lesat system or user emulation is enabled.

Signed-off-by: Carlos Santos <unixmania@gmail.com>
[yann.morin.1998@free.fr:
  - ensure at least one of system, user, or tools
  - guard the machine selection
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>