package/ncurses: Switch to git repo with snapshots
ncurses does not release dot-versions, but the fixes are released as a
set of cumulative patches against the latest version.

So far, we carry the list of the patches against 6.3, and this is a
tedious task to maintain that list (especially as we have a hash for
each of them).

Thomas E. Dickey (the ncurses maintainer, so we can trust him) hosts the
ncurses snapshots on Github, with patches already applied.

So, we switch to using that. This avoids needing to list all incremental
patches, and will make future bumping easier.

Signed-off-by: Dennis Van Hoorick <dennis.vanhoorick@barco.com>
[yann.morin.1998@free.fr:
  - don't use 'v' in filename
  - rework the version string for when there's no snapshot
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>