package/pkg-gemneric: don't add host-ccache dependency to host-{hiredis,pkgconf,zstd}
Now that we made those three package never use ccache to build, we don't
need to add the dependency to host-ccache for those packages.

So far, this is harmless, but when we eventually bump ccache, those
packages will be actual dependency of host-ccache, so we will then
really want to not have them depend on host-ccache.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
[yann.morin.1998@free.fr: split off into its own patch]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>